Fix compilable assets when using a mirrored public folder
When using a mirrored public folder (i.e. `artisan winter:mirror public`), the working directory for the request is set to base_path('public') instead of base_path(), so it is important to use absolute paths when calling File::exists() as it internally just calls file_exists(), which uses the current working directory to resolve relative paths.

Related:
- https://github.com/wintercms/storm/commit/752c16253ca7945e68423e91aa0e8ecbeb4fa064#diff-58c1a5c5e22fc25e75bb02e4af9d2085cfe66b8cfe7c3cffc2f32a4fb60240aaR31
- https://github.com/laravel/framework/issues/45679
- https://github.com/laravel/framework/issues/13243